requirements.txt
streamlit>=1.40.0
openai>=1.0.0
anthropic>=0.20.0
PyMuPDF>=1.23.0
fpdf2>=2.8.0
plotly>=5.18.0
python-dotenv>=1.0.0
